backend/requirements.txt
Django==3.1.3
djangorestframework
djangorestframework-jwt
uritemplate
wheel==0.34.2
psycopg2-binary
django-cors-middleware
django-extensions==1.7.1
drf-yasg==1.20.0
memory_profiler==0.58.0
pytest==6.2.1
pytest-profiling==1.7.0
behave==1.2.6 
behave-django==1.4.0

